add option to control copying of widget bindings
The widget binding js files are copied over to a temporary directory. This is done so that pandoc can later pick up only the widget js file rather than all of its file system siblings. This PR allows the user to control this behavior by introducing a global option htmlwidgets.copybindingdir. It is set to TRUE by default so that current behavior stays unchanged. If a user sets it to FALSE, the widget binding files will not be copied over to a temporary directory and instead served directly from the package directory.