add as.iframe to embed widgets as iframe
This PR allows html widgets to be embedded as iframes. It works by saving the widget to html using saveWidget and then using htmltools to embed it as an iframe.

By default, it saves the widget html to file using selfcontained = TRUE and then uses the srcdoc attribute to embed it as an inline iframe. This allows the resulting html document to be selfcontained. The srcdoc attribute is not supported by IE, and for some reason does not appear to be supported in the Rmd previewer as well.

I have added notes to make it clear that

1. This only works for static Rmds.
2. The iframe wont display in the RStudio viewer when selfcontained = TRUE is chosen.