Move Shiny dependency rendering to be earlier
The previous ordering caused errors when a widget's JS binding's factory
method (or initialize method, for legacy-style bindings) depends on a
dynamic dependency.

You might wonder why a dependency that's used in the intitialize method
would be dynamic and not static. For my case, it's Crosstalk, which
isn't available as a static dependency; you use crosstalk::crosstalkLibs
from R to attach the deps. It would be nice if there was a way to
express this in the widget YAML, you could just point to an R
package::function() or package::variable (though I suspect some hacks
may be needed to get around R CMD check, which wants all Imports to be
used I think?).