Support async operations with new versions of Shiny
Shiny has a new function called createRenderFunction that is designed
to support the creation of render functions (like shinyRenderWidget)
that will work with both traditional and async computations.

This function will be added in the next release of Shiny but there's
not a great reason to pull people to that release just for async
functionality they may or may not use. So we use feature detection
to fallback to the old behavior if the version of Shiny doesn't
support this.