Remove indentation when rending HTML into pandoc
When the html object contains nested tags, the default behaviour is for renderTags to indent the code. This causes an issue when the generated text is parsed by pandoc later on. Pandoc parses the indented block as "Verbatim (code) blocks", escaping it as text in the HTML and wrapping it in <pre><code> tags, The result is a non-functional widget.

To avoid this, set indent = FALSE in the call to renderTags, avoiding this opportunity for pandoc to misinterpret the input file.